Time: 0 ms (100%), Space: 8.5 MB (91.21%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -5,7 +5,7 @@ public:
             return false;
         }
         int org = x;
-        int reversenum = 0;
+        long long reversenum = 0;
         while(x > 0){
             int num = x % 10;
             reversenum  = reversenum * 10 + num;
